requirements.txt
fastapi
uvicorn
pillow
requests
memory-profiler